simplify
This outputs a diy csv format. Be careful when parsing it because it does not escape the "," delimeter when it is printing out all the csv characters.

This change also truncates the non-ascii characters by only outputting up to ordinal code points up to the max u8.

This is a quick 'n dirty tool for assisting in the cryptopals challenge. Earlier changes started heading down the path of extra output and additional robustness but I'm not intending on publishing, promoting, or supporting this as a robust or general tool.